Esra

+++ src/main/java/com/coding/meal_kit/controllers/ReviewController.java
@@ -39,8 +39,6 @@ public class ReviewController {
 
 	@GetMapping("/addReview/{id}")
 	public String reviews(@PathVariable("id") Long id, @ModelAttribute("PostAReview") Review review, Model model) {
-//		model.addAttribute("createR", rService.getAll());
-		
 		model.addAttribute("mealId", mService.getMealbyID(id));
 		return "/meal/review.jsp";
 	}
@@ -60,8 +58,8 @@ public class ReviewController {
 	
 	@DeleteMapping("/delete/{id}/review")
 	public String deleteReview(@PathVariable("id") Long id, HttpSession session, Model  model) {
-//		if (session.getAttribute("userId") == null)
-//			return "redirect:";
+		if (session.getAttribute("userId") == null)
+			return "redirect:";
 		User user = this.uService.findById((Long) session.getAttribute("userId"));
 		rService.deleteReview(id);
 		return "redirect:/profile/" + user.getId();
@@ -69,7 +67,6 @@ public class ReviewController {
 	
 
 	@GetMapping("/edit/{id}/review")
-
 	public String editReview( @ModelAttribute("Edit") Review review, @PathVariable("id") Long id, Model model) {
 		Review rev = rService.getOneR(id);
 		//model.addAttribute("pUser", uService.findById(rev.getReviewedBy().getId()));
